+++ src/main/kotlin/com/jkjamies/cammp/feature/presentationgenerator/data/ModulePackageRepositoryImpl.kt
@@ -1,14 +1,19 @@
 package com.jkjamies.cammp.feature.presentationgenerator.data
 
+import dev.zacsweers.metro.AppScope
 import com.jkjamies.cammp.feature.presentationgenerator.domain.repository.ModulePackageRepository
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VfsUtil
+import dev.zacsweers.metro.ContributesBinding
+import dev.zacsweers.metro.Inject
 import java.nio.file.Path
 
 /**
  * Implementation of [ModulePackageRepository] that uses IntelliJ's Virtual File System (VFS)
  * to find the most appropriate package name within a module.
  */
+@ContributesBinding(AppScope::class)
+@Inject
 class ModulePackageRepositoryImpl : ModulePackageRepository {
     override fun findModulePackage(moduleDir: Path): String? {
         val vf = LocalFileSystem.getInstance().refreshAndFindFileByPath(moduleDir.toString())