+++ src/test/kotlin/com/jkjamies/cammp/feature/usecasegenerator/presentation/UseCaseViewModelTest.kt
@@ -33,9 +33,10 @@ class UseCaseViewModelTest : BehaviorSpec({
     beforeContainer {
         clearAllMocks()
         viewModel = UseCaseViewModel(
+            domainPackage = "",
+            scope = testScope,
             generator = mockGenerator,
-            loadRepositories = mockLoadRepositories,
-            scope = testScope
+            loadRepositories = mockLoadRepositories
         )
     }